+++ src/main/java/com/spotify/dns/ChangeNotifiers.java
@@ -25,10 +25,33 @@ public final class ChangeNotifiers {
   private ChangeNotifiers() {
   }
 
+  /**
+   * Creates a {@link ChangeNotifier} that aggregates the records provided by a list of notifiers.
+   *
+   * A change event on any of the input notifiers will propagate up the the returned notifier.
+   * The set of previous and current records contained in the event will be the union of all
+   * records in the input notifiers, before and after the change event.
+   *
+   * @param notifiers  A list of notifiers to aggregate
+   * @param <T> The record type
+   * @return A notifier with the described behaviour
+   */
   public static <T> ChangeNotifier<T> aggregate(ChangeNotifier<T>... notifiers) {
     return new AggregatingChangeNotifier<T>(Arrays.asList(notifiers));
   }
 
+  /**
+   * Create a {@link ChangeNotifier} with a static set of records.
+   *
+   * This notifier will never generate any change events. Thus any attached
+   * {@link ChangeNotifier.Listener} will at most get one initial call to
+   * {@link ChangeNotifier.Listener#onChange(ChangeNotifier.ChangeNotification)}
+   * if they are attached with the {@code fire} argument set to {@code true}.
+   *
+   * @param records  The records that the notifier will contain
+   * @param <T> The record type
+   * @return A notifier with a static set of records
+   */
   public static <T> ChangeNotifier<T> staticRecords(T... records) {
     return new StaticChangeNotifier<T>(Sets.newHashSet(records));
   }
