+++ src/main/java/com/spotify/dns/DnsSrvResolvers.java
@@ -16,28 +16,16 @@
 
 package com.spotify.dns;
 
-import com.google.common.base.Function;
-import com.google.common.base.Functions;
-import com.google.common.util.concurrent.MoreExecutors;
-import com.google.common.util.concurrent.ThreadFactoryBuilder;
-
 import com.spotify.dns.statistics.DnsReporter;
 
 import org.xbill.DNS.Lookup;
 
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.ScheduledThreadPoolExecutor;
-import java.util.concurrent.TimeUnit;
-
-import static com.google.common.base.Preconditions.checkArgument;
-import static com.google.common.base.Preconditions.checkNotNull;
-import static com.google.common.base.Preconditions.checkState;
 import static com.google.common.primitives.Ints.checkedCast;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static java.util.concurrent.TimeUnit.SECONDS;
 
 /**
- * Provides utility methods for instantiating and working with DnsSrvResolvers.
+ * Provides builders for configuring and instantiating {@link DnsSrvResolver}s.
  */
 public final class DnsSrvResolvers {
 
@@ -45,117 +33,6 @@ public final class DnsSrvResolvers {
 
   public static DnsSrvResolverBuilder newBuilder() {
     return new DnsSrvResolverBuilder();
-  }
-
-  public static DnsSrvWatcherBuilder<LookupResult> newWatcherBuilder(DnsSrvResolver resolver) {
-    checkNotNull(resolver, "resolver");
-
-    return new DnsSrvWatcherBuilder<LookupResult>(resolver, Functions.<LookupResult>identity());
-  }
-
-  public static <T> DnsSrvWatcherBuilder<T> newWatcherBuilder(
-      DnsSrvResolver resolver,
-      Function<LookupResult, T> resultTransformer) {
-
-    checkNotNull(resolver, "resolver");
-    checkNotNull(resultTransformer, "resultTransformer");
-
-    return new DnsSrvWatcherBuilder<T>(resolver, resultTransformer);
-  }
-
-  public static final class DnsSrvWatcherBuilder<T> {
-
-    private final DnsSrvResolver resolver;
-    private final Function<LookupResult, T> resultTransformer;
-
-    private final boolean polling;
-    private final long pollingInterval;
-    private final TimeUnit pollingIntervalUnit;
-
-    private final DnsSrvWatcherFactory<T> dnsSrvWatcherFactory;
-
-    private final ScheduledExecutorService scheduledExecutorService;
-
-    private DnsSrvWatcherBuilder(
-        DnsSrvResolver resolver,
-        Function<LookupResult, T> resultTransformer) {
-      this(resolver, resultTransformer, false, 0, null, null, null);
-    }
-
-    private DnsSrvWatcherBuilder(
-        DnsSrvResolver resolver,
-        Function<LookupResult, T> resultTransformer,
-        boolean polling,
-        long pollingInterval,
-        TimeUnit pollingIntervalUnit,
-        DnsSrvWatcherFactory<T> dnsSrvWatcherFactory,
-        ScheduledExecutorService scheduledExecutorService) {
-      this.resolver = resolver;
-      this.resultTransformer = resultTransformer;
-      this.polling = polling;
-      this.pollingInterval = pollingInterval;
-      this.pollingIntervalUnit = pollingIntervalUnit;
-      this.dnsSrvWatcherFactory = dnsSrvWatcherFactory;
-      this.scheduledExecutorService = scheduledExecutorService;
-    }
-
-    public DnsSrvWatcher<T> build() {
-      checkState(polling ^ dnsSrvWatcherFactory != null, "specify either polling or custom trigger");
-
-      final ChangeNotifierFactory<T> changeNotifierFactory =
-          new ChangeNotifierFactory<T>() {
-            @Override
-            public RunnableChangeNotifier<T> create(String fqdn) {
-              return new ServiceResolvingChangeNotifier<T>(resolver, fqdn, resultTransformer);
-            }
-          };
-
-      DnsSrvWatcherFactory<T> watcherFactory;
-      if (polling) {
-        final ScheduledExecutorService executor =
-            scheduledExecutorService != null
-            ? scheduledExecutorService
-            : MoreExecutors.getExitingScheduledExecutorService(
-                new ScheduledThreadPoolExecutor(
-                    1, new ThreadFactoryBuilder().setNameFormat("dns-lookup-%d").build()),
-                0, SECONDS);
-
-        watcherFactory = new DnsSrvWatcherFactory<T>() {
-          @Override
-          public DnsSrvWatcher<T> create(ChangeNotifierFactory<T> changeNotifierFactory) {
-            return new PollingDnsSrvWatcher<T>(changeNotifierFactory, executor, pollingInterval,
-                                               pollingIntervalUnit);
-          }
-        };
-      } else {
-        watcherFactory = checkNotNull(dnsSrvWatcherFactory);
-      }
-
-      return watcherFactory.create(changeNotifierFactory);
-    }
-
-    public DnsSrvWatcherBuilder<T> polling(long pollingInterval, TimeUnit pollingIntervalUnit) {
-      checkArgument(pollingInterval > 0);
-      checkNotNull(pollingIntervalUnit, "pollingIntervalUnit");
-
-      return new DnsSrvWatcherBuilder<T>(resolver, resultTransformer, true, pollingInterval,
-                                         pollingIntervalUnit, dnsSrvWatcherFactory,
-                                         scheduledExecutorService);
-    }
-
-    public DnsSrvWatcherBuilder<T> usingExecutor(ScheduledExecutorService scheduledExecutorService) {
-      return new DnsSrvWatcherBuilder<T>(resolver, resultTransformer, polling, pollingInterval,
-                                         pollingIntervalUnit, dnsSrvWatcherFactory,
-                                         scheduledExecutorService);
-    }
-
-    public DnsSrvWatcherBuilder<T> customTrigger(DnsSrvWatcherFactory<T> watcherFactory) {
-      checkNotNull(watcherFactory, "watcherFactory");
-
-      return new DnsSrvWatcherBuilder<T>(resolver, resultTransformer, true, pollingInterval,
-                                         pollingIntervalUnit, watcherFactory,
-                                         scheduledExecutorService);
-    }
   }
 
   public static final class DnsSrvResolverBuilder {
